chore: move apps-engine lib to ee
