+++ apps/meteor/app/apps/server/bridges/commands.ts
@@ -5,7 +5,7 @@ import { CommandBridge } from '@rocket.chat/apps-engine/server/bridges/CommandBr
 import type { IMessage, RequiredField, SlashCommand } from '@rocket.chat/core-typings';
 
 import { slashCommands } from '../../../utils/server';
-import { Utilities } from '../../lib/misc/Utilities';
+import { Utilities } from '../../../../ee/lib/misc/Utilities';
 import type { AppServerOrchestrator } from '../../../../ee/server/apps/orchestrator';
 import { parseParameters } from '../../../../lib/utils/parseParameters';
