+++ apps/meteor/app/apps/server/bridges/uploads.ts
@@ -4,7 +4,7 @@ import type { IUpload } from '@rocket.chat/apps-engine/definition/uploads';
 import type { IUploadDetails } from '@rocket.chat/apps-engine/definition/uploads/IUploadDetails';
 
 import { FileUpload } from '../../../file-upload/server';
-import { determineFileType } from '../../lib/misc/determineFileType';
+import { determineFileType } from '../../../../ee/lib/misc/determineFileType';
 import type { AppServerOrchestrator } from '../../../../ee/server/apps/orchestrator';
 
 const getUploadDetails = (details: IUploadDetails): Partial<IUploadDetails> => {
