+++ apps/meteor/app/apps/server/converters/rooms.js
@@ -2,7 +2,7 @@ import { RoomType } from '@rocket.chat/apps-engine/definition/rooms';
 import { LivechatVisitors } from '@rocket.chat/models';
 
 import { Rooms, Users, LivechatDepartment } from '../../../models/server';
-import { transformMappedData } from '../../lib/misc/transformMappedData';
+import { transformMappedData } from '../../../../ee/lib/misc/transformMappedData';
 
 export class AppRoomsConverter {
 	constructor(orch) {
