+++ apps/meteor/app/apps/server/converters/visitors.js
@@ -1,6 +1,6 @@
 import { LivechatVisitors } from '@rocket.chat/models';
 
-import { transformMappedData } from '../../lib/misc/transformMappedData';
+import { transformMappedData } from '../../../../ee/lib/misc/transformMappedData';
 
 // TODO: check if functions from this converter can be async
 export class AppVisitorsConverter {
