+++ apps/meteor/app/ui-message/client/actionButtons/tabbar.ts
@@ -1,7 +1,7 @@
 import type { IUIActionButton } from '@rocket.chat/apps-engine/definition/ui';
 
 import { addAction, deleteAction } from '../../../../client/views/room/lib/Toolbox';
-import { Utilities } from '../../../apps/lib/misc/Utilities';
+import { Utilities } from '../../../../ee/lib/misc/Utilities';
 import { t } from '../../../utils/client';
 import { triggerActionButtonAction } from '../ActionManager';
 import { applyButtonFilters } from './lib/applyButtonFilters';
