+++ apps/meteor/ee/lib/misc/formatAppInstanceForRest.ts
